fix: share setter bodies so coverage instruments one copy
The amalgamation copies of set_username / set_password / set_hash sat
behind ADA_URL_SETTERS_SEPARATE_TU and were compiled out of the
coverage build, which dropped codecov/patch below 60%. One header is
now included from the isolated TU (tests) and from url_aggregator.cpp
(amalgamation).
